refactor(store): split forms, scopes, fields and views out of action.ts (#711)
Completes the split. action.ts is 47 lines and re-exports only; the
2,929-line file is now ten modules, none over 661 lines.

  shared     41    formulas   72    folders   81    fields    249
  schemas   291    agents    294    scopes   305    databases 463
  views     551    forms     661

The three temporary exports from the second cut are gone as predicted:
sortAndRemoveDupSchemas, addNsfDesign and isActiveAgent are now
module-to-module references rather than reaching back into the barrel.
The final graph is exactly the one measured before any code moved:

  shared    <- everything
  schemas   <- scopes
  agents    <- views
  forms     <- fields
  databases <- forms

No cycles, and none possible without adding an edge by hand.

Correction to the two earlier commits in this series: they claim `git
log --follow` still reaches a moved thunk's history. It does not. --follow
tracks renames, and this was a split -- action.ts still exists, so git
sees no rename to follow. Verified: `git log --follow scopes.ts` shows
only the split commit. What does work is a content search,

  git log -S'export const changeScope' -- src/store/databases/

which reaches back to the initial release. That is now documented in the
barrel instead of the false claim.

lint 0 · build 0 · 94 files / 1138 tests, unchanged across all three cuts
